requirements.txt
decorator==4.2.1
hupper==1.0
mysql-replication==0.18
mysqlclient==1.3.12
PasteDeploy==1.5.2
pbr==3.1.1
plaster==1.0
plaster-pastedeploy==0.4.2
PyMySQL==0.8.0
pyramid==1.9.1
pyramid-tm==2.2
repoze.lru==0.7
six==1.11.0
sortedcontainers==1.5.9
SQLAlchemy==1.2.4
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
transaction==2.2.0
translationstring==1.3
venusian==1.1.0
waitress==1.1.0
WebOb==1.8.0rc1
zope.deprecation==4.3.0
zope.interface==4.4.3
zope.sqlalchemy==1.0
